client: Don't delete idtoken until we get a new one
This is in attempt to remove the failure state where the entire network
of displays can be immediately wiped by a misconfigured server. The
primary cause of a misconfigured server is one pointing at an empty or
incorrect database, such that the actual displays do not have
corresponding records.

Keeping the previous token allows a device reboot
to restore functionality; a timer is also implemented to retry the token
regularly to allow self-healing without a manual fleet reboot.